+++ ta4j-core/src/main/java/org/ta4j/core/BaseTimeSeries.java
@@ -64,7 +64,7 @@ public class BaseTimeSeries implements TimeSeries {
     private int removedBarsCount = 0;
     /** True if the current series is constrained (i.e. its indexes cannot change), false otherwise */
     private boolean constrained = false;
-
+    /** Num type function **/
     protected final Function<Number, Num> numFunction;
 
     /**
@@ -440,6 +440,9 @@ public class BaseTimeSeries implements TimeSeries {
         private boolean isConstrained;
         private int maxBarCount;
 
+        /** Default Num type function **/
+        private static Function<Number, Num> defaultFunction = PrecisionNum::valueOf;
+
         public SeriesBuilder(){
             initValues();
         }
@@ -447,7 +450,7 @@ public class BaseTimeSeries implements TimeSeries {
         private void initValues() {
             this.bars = new ArrayList<>();
             this.name = "unnamed_series";
-            this.numFunction = PrecisionNum::valueOf;
+            this.numFunction = SeriesBuilder.defaultFunction;
             this.isConstrained = false;
             this.maxBarCount = Integer.MAX_VALUE;
         }
@@ -507,5 +510,10 @@ public class BaseTimeSeries implements TimeSeries {
             numFunction = PrecisionNum::valueOf;
             return this;
         }
+
+        public static void setDefaultFunction(Function<Number, Num> defaultFunction) {
+            SeriesBuilder.defaultFunction = defaultFunction;
+        }
+
     }
 }